made project pip-installable
By providing `hopla = hopla:entry_cmd` we can now
run hopla directly from the cmdline after a pip install.

Also removed old shell deps, now we have less python
code in __init__.py
