--- hopla/feed/pet.help
@@ -1,12 +1,0 @@
-hopla feed pet - feed a pet
-
-
-usage:
-    hopla feed pet --help
-    hopla feed pet {pet_name} {food_name}
-
-
-Example:
-```bash
-hopla feed pet Beetle-Skeleton Fish
-```
