+++ setup.cfg
@@ -1,13 +1,18 @@
 [metadata]
 name = hopla
-version = 0.0.1-alpha
+version = 0.0.6-alpha
 
 [options]
-packages = find:
+package_dir =
+    =src
+packages = find_namespace:
 install_requires =
     requests
     Click
 
+[options.packages.find]
+where = src
+
 [options.entry_points]
 console_scripts =
-    hopla = main:hopla
+    hopla = hopla:entry_cmd
