+++ src/hopla/__main__.py
@@ -1,0 +1,5 @@
+#!/usr/bin/env python3
+from . import entry_cmd
+
+if __name__ == "__main__":
+    entry_cmd()
