+++ src/hopla/complete/bash.sh
