+++ src/hopla/get/user-info.help
