+++ src/hopla/subgroups/get.help
